Remove configuration of securityConfig.filter.loginUrl; it doesn't work properly in all cases, there are other options available, and it may not even be needed.
Specifically, if you're deployed as a war (such as via run-war or an external servlet container) and grails.serverUrl isn't configured, grailsLinkGenerator won't provide a serverBaseUrl outside of a request scope.

In this case, we won't be able to assemble a login URL.  This setting on the filter doesn't appear to be needed by grails-shiro-cas, as we explicitly get our login url from our own config.
For other filters, if they need this information, the user can configure security.shiro.filter.loginUrl as documented by the grails-shiro plugin.  Alternatively, they too can delegate to our config logic.