Fix ticket validation with new multi-domain approach
In the case of ticket validation in a shiro realm, GrailsWebRequest.lookup() returns null, causing the default link generator logic to not work as intended.
However, in this case we have an additional bit of information: the current request is the one that we're attempting to validate.
Thus, we can default to the current URL.